+++ CHANGELOG.rst
@@ -17,6 +17,7 @@ Quetzal version numbers follow `semantic versioning <http://semver.org>`_.
 * Added Sphinx documentation structure.
 * Added file delete endpoint.
 * Added file state in base metadata.
+* Added a simple/naive workspace commit conflict detection.
 
 0.1.0 (2019-03-05)
 ------------------